third_party/blink/web_tests/virtual/font-variant-emoji/external/wpt/css/css-fonts/font-shorthand-serialization-prevention-expected.txt

This is a testharness.js-based test.
Found 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Setting font-variant-emoji to initial should prevent the font shorthand from serializing in specified style
  assert_equals: expected "" but got "16px serif"
[FAIL] Setting font-variant-emoji to inherit should prevent the font shorthand from serializing in specified style
  assert_equals: expected "" but got "16px serif"
[FAIL] Setting font-variant-emoji to unset should prevent the font shorthand from serializing in specified style
  assert_equals: expected "" but got "16px serif"
[FAIL] Setting font-variant-emoji to revert should prevent the font shorthand from serializing in specified style
  assert_equals: expected "" but got "16px serif"
[FAIL] Setting font-variant-emoji to revert-layer should prevent the font shorthand from serializing in specified style
  assert_equals: expected "" but got "16px serif"
[FAIL] Setting font-variant-emoji to text should prevent the font shorthand from serializing in specified style
  assert_equals: expected "" but got "16px serif"
[FAIL] Setting font-variant-emoji to text should prevent the font shorthand from serializing in computed style
  assert_equals: expected "" but got "16px serif"
Harness: the test ran to completion.
